+++ tests/test_api.py
@@ -16,7 +16,7 @@ DATABASE_URL = os.environ["DATABASE_URL"]
 
 
 class Transaction(IdModel):
-    date: dt.date = None
+    date: dt.date = None  # type: ignore
     price: int
 
     @validator("date", pre=True, always=True)
@@ -78,7 +78,7 @@ async def transactions(repo: TransactionRepo) -> List[Transaction]:
 
 
 async def test_base_repository_insert_sets_id_and_inserts_to_db(
-    repo: TransactionRepo
+        repo: TransactionRepo
 ) -> None:
     trans = Transaction(price=100)
 
@@ -87,6 +87,7 @@ async def test_base_repository_insert_sets_id_and_inserts_to_db(
     assert trans.id == 1
 
     db_trans = await repo.first()
+    assert db_trans
     assert db_trans.id == trans.id
 
 
@@ -108,20 +109,22 @@ async def test_base_repo_update_updates_row_in_db(repo: TransactionRepo) -> None
     await repo.update(trans)
 
     updated_trans = await repo.first()
+    assert updated_trans
     assert updated_trans.price == trans.price
     assert updated_trans.date == trans.date
 
 
 async def test_base_repo_first_return_first_matching_row(
-    repo: TransactionRepo, transactions: List[Transaction]
+        repo: TransactionRepo, transactions: List[Transaction]
 ) -> None:
     trans = await repo.first(transactions_table.c.price == 100)
 
+    assert trans
     assert trans.id == transactions[0].id
 
 
 async def test_base_repo_get_all_return_all_rows_filtered_and_sorted(
-    repo: TransactionRepo, transactions: List[Transaction]
+        repo: TransactionRepo, transactions: List[Transaction]
 ) -> None:
     db_transactions = await repo.get_all(
         filters=[transactions_table.c.price == 100], orders=[transactions_table.c.date]
@@ -130,7 +133,7 @@ async def test_base_repo_get_all_return_all_rows_filtered_and_sorted(
 
 
 async def test_base_repo_delete_deletes_row_from_db(
-    repo: TransactionRepo, transactions: List[Transaction]
+        repo: TransactionRepo, transactions: List[Transaction]
 ) -> None:
     await repo.delete(transactions_table.c.price == 100)
 
@@ -139,7 +142,7 @@ async def test_base_repo_delete_deletes_row_from_db(
 
 
 async def test_transaction_repo_custom_method_works(
-    repo: TransactionRepo, transactions: List[Transaction]
+        repo: TransactionRepo, transactions: List[Transaction]
 ) -> None:
     sum_ = await repo.sum()
 
@@ -147,14 +150,16 @@ async def test_transaction_repo_custom_method_works(
 
 
 async def test_base_repo_get_by_id_returns_row_with_id(
-    repo: TransactionRepo, transactions: List[Transaction]
+        repo: TransactionRepo, transactions: List[Transaction]
 ) -> None:
-    db_trans = await repo.get_by_id(transactions[0].id)
+    transaction_id = transactions[0].id
+    assert transaction_id
+    db_trans = await repo.get_by_id(transaction_id)
     assert db_trans == transactions[0]
 
 
 async def test_base_repo_get_or_create_creates_entity_if_no_entities(
-    repo: TransactionRepo
+        repo: TransactionRepo
 ) -> None:
     price = 400
     trans, created = await repo.get_or_create(defaults={"price": price})
@@ -163,7 +168,7 @@ async def test_base_repo_get_or_create_creates_entity_if_no_entities(
 
 
 async def test_base_repo_get_or_create_returns_entity_if_match(
-    repo: TransactionRepo, transactions: List[Transaction]
+        repo: TransactionRepo, transactions: List[Transaction]
 ) -> None:
     price = 400
     trans, created = await repo.get_or_create(