+++ CHANGELOG.md
@@ -8,7 +8,9 @@
 - repka.api.BaseRepository.get_or_create - get first entity by filters or create new with defaults
 - repka.api.BaseRepository.serializer property for defining serialization behaviour
 - repka.api.BaseRepository.deserializer property for defining deserialization behaviour
+
 ### Removed
+
 - repka.api.BaseRepository.entity_type property. The same functionality can be provided by deserializer property
 
 ## 0.1.0 - 2019-07-02